+++ internal/ui/ui.go
@@ -27,6 +27,8 @@ type backend interface {
 	CreateFolder(uid, name, parent string) (doc *storage.Document, err error)
 	UpdateDocument(uid, docID, name, parent string) (err error)
 	UpdateRmDoc(uid, docID string, stream io.Reader) (err error)
+	GetReadingProgress(uid, docID string) (*viewmodel.ReadingProgress, error)
+	UpdateReadingProgress(uid, docID string, page int) error
 	DeleteDocument(uid, docID string) (err error)
 	Sync(uid string)
 }
@@ -49,6 +51,7 @@ type blobHandler interface {
 	CreateBlobDocument(uid, name, parent string, reader io.Reader) (doc *storage.Document, err error)
 	UpdateBlobDocument(uid, docID, name, parent string) (err error)
 	UpdateBlobDocumentFromRmDoc(uid, docID string, reader io.Reader) (err error)
+	UpdateBlobDocumentReadingPosition(uid, docID string, page int) error
 	DeleteBlobDocument(uid, docID string) (err error)
 	CreateBlobFolder(uid, name, parent string) (doc *storage.Document, err error)
 	Export(uid, docid string) (io.ReadCloser, error)